Fix lint warnings and move images to image folder
